changing job repo to be interface
